Initial commit
Set up project structure and add initial files.

Finish revising the login and account creation
