+++ resources/js/Pages/Account Creation/components/Step3GameDetails2.jsx
@@ -1,7 +1,6 @@
 // Step3GameDetails.jsx
-import React, { useState, useRef, useEffect } from "react";
+import React, { useState, useRef } from "react";
 import MLLogin from '../../MLLoginApi/MLLogin.jsx';
-import styles from '../register.module.scss'; 
 
 
 const Step3GameDetails = ({
@@ -17,35 +16,7 @@ const Step3GameDetails = ({
   const [isMlIdValid, setIsMlIdValid] = useState(false);
   const [userInfo, setUserInfo] = useState(null);
 
-
-  // Add state to hold your dropdown options
-  const [dropdownOptions, setDropdownOptions] = useState({
-    rankings: [],
-    in_game_roles: [],
-    main_heroes: []
-  });
-
   const loginRef = useRef();
-
-
-   // Use useEffect to fetch your JSON data
-  useEffect(() => {
-    fetch('/json/lists.json')
-      .then(response => {
-        if (!response.ok) {
-          throw new Error(`HTTP error! status: ${response.status}`);
-        }
-        return response.json();
-      })
-      .then(data => {
-        setDropdownOptions(data); // Set the fetched options to state
-      })
-      .catch(error => {
-        console.error("Error fetching dropdown lists:", error);
-        // You might want to set an error message or provide fallback options
-      });
-  }, []); // Empty dependency array means this runs once on component mount
-
 
   const handleLoginInfo = (info) => {
     console.log('Received Login Info:', info);
@@ -163,13 +134,9 @@ const Step3GameDetails = ({
   };
 
   return (
-   <div className={`${styles['step3-outer-wrapper']} flex flex-col items-center`}>
-      <h1 className={`${styles['title-register']} text-white mb-2 text-2xl md:text-[2.5rem]`}>
-        CREATE MSL ACCOUNT
-      </h1>
-      <h2 className={`${styles['subtitle-register']} text-white`}>
-        MLBB DETAILS
-      </h2>
+    <div className="">
+      <h1 className="title-register">CREATE MSL ACCOUNT</h1>
+      <h2 className="subtitle-register">MLBB DETAILS</h2>
 
       {/* Dynamic Progress Bar for Step 3 */}
       {(() => {
@@ -186,318 +153,373 @@ const Step3GameDetails = ({
         const percent = 51 + Math.round((filled / requiredFields.length) * (75 - 51));
 
         return (
-          <div className="my-4 w-full max-w-2xl px-4">
-            <div className="h-3 bg-gray-200 rounded-full overflow-hidden mb-1">
-              <div
-                className="h-full bg-yellow-400 transition-all duration-300"
-                style={{ width: `${percent}%` }}
-              />
+          <div style={{ margin: "16px 0" }}>
+            <div style={{ height: "12px", background: "#eee", borderRadius: "6px", overflow: "hidden", marginBottom: "4px" }}>
+              <div style={{ width: `${percent}%`, height: "100%", background: "#f1c40f", transition: "width 0.3s" }} />
             </div>
-            <div className="text-xs text-gray-100 text-right">
+            <div style={{ fontSize: "12px", color: "#555" }}>
               Step 3 of 4 &mdash; {percent}% of this step complete
             </div>
           </div>
         );
-        })()}
-       <div className="w-full max-w-2xl relative"> {/* Adjusted container */}
+      })()}
+      <div className="form-register-container">
         {!isBlurred ?
-          <div className={`${styles.blurred} absolute inset-0 flex flex-col md:flex-row justify-center items-center w-full h-full bg-black bg-opacity-50 rounded-lg z-20 p-4`}>
-            <img src="/mlbb-logo.png" alt="MSL Logo" className="w-36 md:w-48 mb-4 md:mb-0 md:mr-8" />
-            <div className={`${styles.loading} scale-75 md:scale-100 rotate-90 md:rotate-0 mb-4 md:mb-0 md:mr-8`}> {/* Apply the SCSS loading class */}
-              <svg width="64px" height="48px" viewBox="0 0 64 48">
-                <polyline
-                  points="0.157 23.954, 14 23.954, 21.843 48, 43 0, 50 24, 64 24"
-                  id="back"
-                  // Removed Tailwind classes here to let SCSS manage stroke
-                />
-                <polyline
-                  points="0.157 23.954, 14 23.954, 21.843 48, 43 0, 50 24, 64 24"
-                  id="front"
-                  // Removed Tailwind classes here to let SCSS manage stroke
-                />
-              </svg>
-            </div>
-            <div className={styles['electric-wrapper']}> {/* Apply the SCSS electric-wrapper class */}
-              <div onClick={() => {
-                handleLoginClick()
-                }}
-                className={`${styles['neon-button']} text-sm md:text-base px-4 md:px-8 py-2 md:py-3 flex items-center justify-center gap-2`}> {/* Apply the SCSS neon-button class */}
-                {isLoading ? (
-                  <>
-                    <svg className="animate-spin h-5 w-5 text-white" xmlns="http://www.w3.org/2000/svg" fill="none" viewBox="0 0 24 24">
-                      <circle className="opacity-25" cx="12" cy="12" r="10" stroke="currentColor" strokeWidth="4"></circle>
-                      <path className="opacity-75" fill="currentColor" d="M4 12a8 8 0 018-8V0C5.373 0 0 5.373 0 12h4zm2 5.291A7.962 7.962 0 014 12H0c0 3.042 1.135 5.824 3 7.938l3-2.647z"></path>
-                    </svg>
-                    <span>Connecting...</span>
-                  </>
-                ) : (
-                  <span>Connect</span>
-                )}
-                <div className={styles.lightning}></div> {/* Restored the lightning bolt element */}
+          <div className="blurred">
+            <div className="flex flex-col md:flex-row justify-center items-center w-full gap-4 h-full pb-10 px-4">
+              <img src="/mlbb-logo.png" alt="MSL Logo" className="msl-logo w-48 md:w-48" />
+              <div className="loading scale-75 md:scale-100 rotate-90 md:rotate-0">
+                <svg width="64px" height="48px" viewBox="0 0 64 48">
+                  <polyline
+                    points="0.157 23.954, 14 23.954, 21.843 48, 43 0, 50 24, 64 24"
+                    id="back"
+                  />
+                  <polyline
+                    points="0.157 23.954, 14 23.954, 21.843 48, 43 0, 50 24, 64 24"
+                    id="front"
+                  />
+                </svg>
               </div>
-              <MLLogin
-                ref={loginRef}
-                onLoginInfo={handleLoginInfo}
-                onLoginClose={() => {
-                  setIsLoading(false);
+              <div className="electric-wrapper">
+                <div onClick={() => {
+                  handleLoginClick()
+                  }}
+                  className="neon-button text-sm md:text-base px-4 md:px-8 py-2 md:py-3 flex items-center justify-center gap-2">
+                  {isLoading ? (
+                    <>
+                      <svg className="animate-spin h-5 w-5 text-white" xmlns="http://www.w3.org/2000/svg" fill="none" viewBox="0 0 24 24">
+                        <circle className="opacity-25" cx="12" cy="12" r="10" stroke="currentColor" strokeWidth="4"></circle>
+                        <path className="opacity-75" fill="currentColor" d="M4 12a8 8 0 018-8V0C5.373 0 0 5.373 0 12h4zm2 5.291A7.962 7.962 0 014 12H0c0 3.042 1.135 5.824 3 7.938l3-2.647z"></path>
+                      </svg>
+                      <span>Connecting...</span>
+                    </>
+                  ) : (
+                    <span>Connect</span>
+                  )}
+                </div>
+                <MLLogin
+                  ref={loginRef}
+                  onLoginInfo={handleLoginInfo}
+                  onLoginClose={() => {
+                    setIsLoading(false);
+                  }}
+                  onLoginSuccess={() => {
+                    setIsLoading(false);
+                    setIsBlurred(true);
+                  }}
+                />
+              </div>
+            </div>
+          </div>
+        : null}
+        <div className={`form-register`}>
+          {isCheckingMlId && (
+            <div className="mb-4 p-3 bg-blue-100 border border-blue-400 text-blue-700 rounded z-10">
+              <div className="flex items-center">
+                <svg className="animate-spin h-4 w-4 mr-2" xmlns="http://www.w3.org/2000/svg" fill="none" viewBox="0 0 24 24">
+                  <circle className="opacity-25" cx="12" cy="12" r="10" stroke="currentColor" strokeWidth="4"></circle>
+                  <path className="opacity-75" fill="currentColor" d="M4 12a8 8 0 018-8V0C5.373 0 0 5.373 0 12h4zm2 5.291A7.962 7.962 0 014 12H0c0 3.042 1.135 5.824 3 7.938l3-2.647z"></path>
+                </svg>
+                Checking ML ID availability...
+              </div>
+            </div>
+          )}
+          
+          {localError && (
+            <div className="mb-4 p-3 bg-red-100 border border-red-400 text-red-700 rounded z-10">
+              {localError}
+            </div>
+          )}
+          
+          {isMlIdValid && !isCheckingMlId && (
+            <div className="mb-4 p-3 bg-green-100 border border-green-400 text-green-700 rounded">
+              <div className="flex items-center">
+                <svg className="h-4 w-4 mr-2" fill="currentColor" viewBox="0 0 20 20">
+                  <path fillRule="evenodd" d="M10 18a8 8 0 100-16 8 8 0 000 16zm3.707-9.293a1 1 0 00-1.414-1.414L9 10.586 7.707 9.293a1 1 0 00-1.414 1.414l2 2a1 1 0 001.414 0l4-4z" clipRule="evenodd" />
+                </svg>
+                ML Account is available! Please complete the form below.
+              </div>
+            </div>
+          )}
+          
+          <div className="form-row-register">
+            <div className="input-group-register left-side-register">
+              <label htmlFor="userId" className="label-register">
+                User ID<span className="required"> *</span>
+              </label>
+              <input
+                type="text"
+                id="userId"
+                name="userId"
+                inputMode="numeric"
+                pattern="\d*"
+                value={formData.userId}
+                onChange={handleInputChange}
+                className="input-field-register"
+                placeholder="e.g. 7373793739"
+                required
+                onKeyPress={e => {
+                  if (!/[0-9]/.test(e.key)) e.preventDefault();
                 }}
-                onLoginSuccess={() => {
-                  setIsLoading(false);
-                  setIsBlurred(true);
+                onPaste={e => {
+                  const pasted = e.clipboardData.getData('Text');
+                  if (!/^\d+$/.test(pasted)) e.preventDefault();
                 }}
+                onBlur={handleAnyInputBlur}
+              disabled/>
+            </div>
+            <div className="input-group-register right-side-register">
+              <label htmlFor="serverId" className="label-register">
+                MLBB Server ID<span className="required"> *</span>
+              </label>
+              <input
+                type="text"
+                id="serverId"
+                name="serverId"
+                inputMode="numeric"
+                pattern="\d*"
+                value={formData.serverId}
+                onChange={handleInputChange}
+                className="input-field-register"
+                placeholder="e.g. 12345"
+                required
+                onKeyPress={e => {
+                  if (!/[0-9]/.test(e.key)) e.preventDefault();
+                }}
+                onPaste={e => {
+                  const pasted = e.clipboardData.getData('Text');
+                  if (!/^\d+$/.test(pasted)) e.preventDefault();
+                }}
+                onBlur={handleAnyInputBlur}
+              disabled/>
+            </div>
+          </div>
+          <div className="form-row-register">
+            <div className="input-group-register left-side-register">
+              <label htmlFor="ign" className="label-register">
+                In-Game Name
+              </label>
+              <input
+                type="text"
+                id="ign"
+                name="ign"
+                value={formData.ign}
+                onChange={handleInputChange}
+                className="input-field-register"
+
+              disabled/>
+            </div>
+            <div className="input-group-register right-side-register">
+              <label htmlFor="squadName" className="label-register">
+                MLBB Squad Name
+              </label>
+              <input
+                type="text"
+                id="squadName"
+                name="squadName"
+                value={formData.squadName}
+                onChange={handleInputChange}
+                className="input-field-register"
+                placeholder="e.g. Legends Squad"
               />
             </div>
           </div>
-        : null }
-          <div className="space-y-4"> {/* Main form container, previously form-register */}
-            {isCheckingMlId && (
-              <div className="mb-4 p-3 bg-blue-100 border border-blue-400 text-blue-700 rounded z-10">
-                <div className="flex items-center">
-                  <svg className="animate-spin h-4 w-4 mr-2" xmlns="http://www.w3.org/2000/svg" fill="none" viewBox="0 0 24 24">
-                    <circle className="opacity-25" cx="12" cy="12" r="10" stroke="currentColor" strokeWidth="4"></circle>
-                    <path className="opacity-75" fill="currentColor" d="M4 12a8 8 0 018-8V0C5.373 0 0 5.373 0 12h4zm2 5.291A7.962 7.962 0 014 12H0c0 3.042 1.135 5.824 3 7.938l3-2.647z"></path>
-                  </svg>
-                  Checking ML ID availability...
-                </div>
-              </div>
-            )}
-            
-            {/* {localError && (
-              <div className="mb-4 p-3 bg-red-700 border border-red-500 text-white rounded z-10">
-                {localError}
-              </div>
-            )} */}
-            
-            {isMlIdValid && !isCheckingMlId && (
-              <div className="mb-4 p-3 bg-green-700 border border-green-500 text-white rounded">
-                <div className="flex items-center">
-                  <svg className="h-4 w-4 mr-2" fill="currentColor" viewBox="0 0 20 20">
-                    <path fillRule="evenodd" d="M10 18a8 8 0 100-16 8 8 0 000 16zm3.707-9.293a1 1 0 00-1.414-1.414L9 10.586 7.707 9.293a1 1 0 00-1.414 1.414l2 2a1 1 0 001.414 0l4-4z" clipRule="evenodd" />
-                  </svg>
-                  ML Account is available! Please complete the form below.
-                </div>
-              </div>
-            )}
-            
-<div className="flex flex-col md:flex-row gap-4">
-  {/* form-row-register */}
-  <div className="flex-1">
-    {/* input-group-register left-side-register */}
-    <label htmlFor="userId" className="block text-white text-sm font-bold mb-2">
-      User ID<span className="text-red-500"> *</span>
-    </label>
-    <input
-      type="text"
-      id="userId"
-      name="userId"
-      inputMode="numeric"
-      pattern="\d*"
-      value={formData.userId}
-      onChange={handleInputChange}
-      className={`${styles['input-field-register']} w-full p-3 text-white border border-gray-700 bg-gray-900 bg-opacity-70 rounded-lg text-base placeholder-gray-500 focus:outline-none focus:border-yellow-400 disabled:opacity-50`}
-      placeholder="e.g. 7373793739"
-      required
-      onKeyPress={(e) => {
-        if (!/[0-9]/.test(e.key)) e.preventDefault();
-      }}
-      onPaste={(e) => {
-        const pasted = e.clipboardData.getData("Text");
-        if (!/^\d+$/.test(pasted)) e.preventDefault();
-      }}
-      onBlur={handleAnyInputBlur}
-      disabled
-    />
-  </div>
-  <div className="flex-1">
-    {/* input-group-register right-side-register */}
-    <label
-      htmlFor="serverId"
-      className="block text-white text-sm font-bold mb-2"
-    >
-      MLBB Server ID<span className="text-red-500"> *</span>
-    </label>
-    <input
-      type="text"
-      id="serverId"
-      name="serverId"
-      inputMode="numeric"
-      pattern="\d*"
-      value={formData.serverId}
-      onChange={handleInputChange}
-      className={`${styles['input-field-register']} w-full p-3 text-white border border-gray-700 bg-gray-900 bg-opacity-70 rounded-lg text-base placeholder-gray-500 focus:outline-none focus:border-yellow-400 disabled:opacity-50`}
-      placeholder="e.g. 12345"
-      required
-      onKeyPress={(e) => {
-        if (!/[0-9]/.test(e.key)) e.preventDefault();
-      }}
-      onPaste={(e) => {
-        const pasted = e.clipboardData.getData("Text");
-        if (!/^\d+$/.test(pasted)) e.preventDefault();
-      }}
-      onBlur={handleAnyInputBlur}
-      disabled
-    />
-  </div>
-</div>
-<div className="flex flex-col md:flex-row gap-4">
-  {/* form-row-register */}
-  <div className="flex-1">
-    {/* input-group-register left-side-register */}
-    <label htmlFor="ign" className="block text-white text-sm font-bold mb-2">
-      In-Game Name
-    </label>
-    <input
-      type="text"
-      id="ign"
-      name="ign"
-      value={formData.ign}
-      onChange={handleInputChange}
-      className={`${styles['input-field-register']} w-full p-3 text-white border border-gray-700 bg-gray-900 bg-opacity-70 rounded-lg text-base placeholder-gray-500 focus:outline-none focus:border-yellow-400 disabled:opacity-50`}
-      disabled
-    />
-  </div>
-  <div className="flex-1">
-    {/* input-group-register right-side-register */}
-    <label
-      htmlFor="squadName"
-      className="block text-white text-sm font-bold mb-2"
-    >
-      MLBB Squad Name
-    </label>
-    <input
-      type="text"
-      id="squadName"
-      name="squadName"
-      value={formData.squadName}
-      onChange={handleInputChange}
-      className={`${styles['input-field-register']} w-full p-3 text-white border border-gray-700 bg-gray-900 bg-opacity-70 rounded-lg text-base placeholder-gray-500 focus:outline-none focus:border-yellow-400`}
-      placeholder="e.g. Legends Squad"
-    />
-  </div>
-</div>
-<div className="flex flex-col md:flex-row gap-4">
-  {/* form-row-register */}
-  <div className="flex-1">
-    {/* input-group-register left-side-register */}
-    <label
-      htmlFor="squadAbbreviation"
-      className="block text-white text-sm font-bold mb-2"
-    >
-      Squad Abbreviation
-    </label>
-    <input
-      type="text"
-      id="squadAbbreviation"
-      name="squadAbbreviation"
-      value={formData.squadAbbreviation}
-      onChange={handleInputChange}
-      className={`${styles['input-field-register']} w-full p-3 text-white border border-gray-700 bg-gray-900 bg-opacity-70 rounded-lg text-base placeholder-gray-500 focus:outline-none focus:border-yellow-400`}
-      placeholder="e.g. El Fili Esports"
-    />
-  </div>
-  <div className="flex-1">
-    {/* input-group-register right-side-register */}
-    <label htmlFor="rank" className="block text-white text-sm font-bold mb-2">
-      Current Rank<span className="text-red-500"> *</span>
-    </label>
-    <select
-          id="rank"
-          name="rank"
-          value={formData.rank}
-          onChange={handleInputChange}
-          className={`${styles['input-field-register']} w-full p-3 text-white border border-gray-700 bg-gray-900 bg-opacity-70 rounded-lg text-base placeholder-gray-500 focus:outline-none focus:border-yellow-400 appearance-none`}
-          style={{
-            backgroundImage: `url("data:image/svg+xml,%3csvg xmlns='http://www.w3.org/2000/svg' viewBox='0 0 16 16'%3e%3cpath fill='none' stroke='%23ffffff' stroke-linecap='round' stroke-linejoin='round' stroke-width='2' d='m2 5 6 6 6-6'/%3e%3c/svg%3e")`,
-            backgroundRepeat: 'no-repeat',
-            backgroundPosition: 'right 0.75rem center',
-            backgroundSize: '1em 1em',
-          }}
-          required
-          onBlur={handleAnyInputBlur}
-        >
-          <option value="" disabled>
-            Select Ranking
-          </option>
-          {dropdownOptions.rankings.map((option, index) => (
-            <option key={index} value={option.value}>
-              {option.text}
-            </option>
-          ))}
-        </select>
-  </div>
-</div>
-<div className="flex flex-col md:flex-row gap-4">
-  {/* form-row-register */}
-  <div className="flex-1">
-    {/* input-group-register left-side-register */}
-    <label
-      htmlFor="inGameRole"
-      className="block text-white text-sm font-bold mb-2"
-    >
-      In-Game Role<span className="text-red-500"> *</span>
-    </label>
-    <select
-      id="inGameRole"
-      name="inGameRole"
-      value={formData.inGameRole}
-      onChange={handleInputChange}
-      className={`${styles['input-field-register']} w-full p-3 text-white border border-gray-700 bg-gray-900 bg-opacity-70 rounded-lg text-base placeholder-gray-500 focus:outline-none focus:border-yellow-400 appearance-none`}
-      style={{
-        backgroundImage: `url("data:image/svg+xml,%3csvg xmlns='http://www.w3.org/2000/svg' viewBox='0 0 16 16'%3e%3cpath fill='none' stroke='%23ffffff' stroke-linecap='round' stroke-linejoin='round' stroke-width='2' d='m2 5 6 6 6-6'/%3e%3c/svg%3e")`,
-        backgroundRepeat: 'no-repeat', /* Add this! */
-        backgroundPosition: 'right 0.75rem center', /* Add this! */
-        backgroundSize: '1em 1em', /* Add this! */
-      }}
-      required
-      onBlur={handleAnyInputBlur}
-    >
-      <option value="" disabled>
-        Select Role
-      </option>
-      {dropdownOptions.in_game_roles.map((option, index) => (
-        <option key={index} value={option.value}>
-          {option.text}
-        </option>
-      ))}
-    </select>
-  </div>
-  <div className="flex-1">
-    {/* input-group-register right-side-register */}
-    <label
-      htmlFor="mainHero"
-      className="block text-white text-sm font-bold mb-2"
-    >
-      Main Hero<span className="text-red-500"> *</span>
-    </label>
-    <select
-      id="mainHero"
-      name="mainHero"
-      value={formData.mainHero}
-      onChange={handleInputChange}
-      className={`${styles['input-field-register']} w-full p-3 text-white border border-gray-700 bg-gray-900 bg-opacity-70 rounded-lg text-base placeholder-gray-500 focus:outline-none focus:border-yellow-400 appearance-none`}
-      style={{
-        backgroundImage: `url("data:image/svg+xml,%3csvg xmlns='http://www.w3.org/2000/svg' viewBox='0 0 16 16'%3e%3cpath fill='none' stroke='%23ffffff' stroke-linecap='round' stroke-linejoin='round' stroke-width='2' d='m2 5 6 6 6-6'/%3e%3c/svg%3e")`,
-        backgroundRepeat: 'no-repeat', /* Crucial addition */
-        backgroundPosition: 'right 0.75rem center', /* Crucial addition */
-        backgroundSize: '1em 1em', /* Crucial addition */
-      }}
-      required
-      onBlur={handleAnyInputBlur}
-    >
-      <option value="" disabled>
-        Select Hero
-      </option>
-      {dropdownOptions.main_heroes.map((option, index) => (
-        <option key={index} value={option.value}>
-          {option.text}
-        </option>
-      ))}
-    </select>
-  </div>
-</div>
+          <div className="form-row-register">
+            <div className="input-group-register left-side-register">
+              <label htmlFor="squadAbbreviation" className="label-register">
+                Squad Abbreviation
+              </label>
+              <input
+                type="text"
+                id="squadAbbreviation"
+                name="squadAbbreviation"
+                value={formData.squadAbbreviation}
+                onChange={handleInputChange}
+                className="input-field-register"
+                placeholder="e.g. El Fili Esports"
+              />
+            </div>
+            <div className="input-group-register right-side-register">
+              <label htmlFor="rank" className="label-register">
+                Current Rank<span className="required"> *</span>
+              </label>
+              <select
+                id="rank"
+                name="rank"
+                value={formData.rank}
+                onChange={handleInputChange}
+                className="input-field-register current-rank-select"
+                required
+                onBlur={handleAnyInputBlur}
+              >
+                <option value="" disabled>Select Ranking</option>
+                <option value="Warrior">Warrior</option>
+                <option value="Elite">Elite</option>
+                <option value="Master">Master</option>
+                <option value="Grandmaster">Grandmaster</option>
+                <option value="Epic">Epic</option>
+                <option value="Legend">Legend</option>
+                <option value="Mythic">Mythic</option>
+                <option value="Mythical Glory">Mythical Honor</option>
+                <option value="Mythical Glory">Mythical Glory</option>
+                <option value="Mythical Glory">Mythical Immortal</option>
+              </select>
+            </div>
+          </div>
+          <div className="form-row-register">
+            <div className="input-group-register left-side-register">
+              <label htmlFor="inGameRole" className="label-register">
+                In-Game Role<span className="required"> *</span>
+              </label>
+              <select
+                id="inGameRole"
+                name="inGameRole"
+                value={formData.inGameRole}
+                onChange={handleInputChange}
+                className="input-field-register in-game-role-select"
+                required
+                onBlur={handleAnyInputBlur}
+              >
+                <option value="" disabled>Select Role</option>
+                <option value="Tank">Tank</option>
+                <option value="Support">Support</option>
+                <option value="Marksman">Marksman</option>
+                <option value="Mage">Mage</option>
+                <option value="Assassin">Assassin</option>
+                <option value="Fighter">Fighter</option>
+              </select>
+            </div>
+            <div className="input-group-register right-side-register">
+              <label htmlFor="mainHero" className="label-register">
+                Main Hero<span className="required"> *</span>
+              </label>
+              <select
+                id="mainHero"
+                name="mainHero"
+                value={formData.mainHero}
+                onChange={handleInputChange}
+                className="input-field-register main-hero-select"
+                required
+                onBlur={handleAnyInputBlur}
+              >
+                <option value="" disabled>Select Hero</option>
+                <option value="Aamon">Aamon</option>
+                <option value="Akai">Akai</option>
+                <option value="Aldous">Aldous</option>
+                <option value="Alice">Alice</option>
+                <option value="Alpha">Alpha</option>
+                <option value="Alucard">Alucard</option>
+                <option value="Angela">Angela</option>
+                <option value="Argus">Argus</option>
+                <option value="Arlott">Arlott</option>
+                <option value="Atlas">Atlas</option>
+                <option value="Aurora">Aurora</option>
+                <option value="Badang">Badang</option>
+                <option value="Balmond">Balmond</option>
+                <option value="Barats">Barats</option>
+                <option value="Baxia">Baxia</option>
+                <option value="Beatrix">Beatrix</option>
+                <option value="Belerick">Belerick</option>
+                <option value="Benedetta">Benedetta</option>
+                <option value="Brody">Brody</option>
+                <option value="Bruno">Bruno</option>
+                <option value="Carmilla">Carmilla</option>
+                <option value="Cecilion">Cecilion</option>
+                <option value="Chang'e">Chang'e</option>
+                <option value="Chou">Chou</option>
+                <option value="Clint">Clint</option>
+                <option value="Cyclops">Cyclops</option>
+                <option value="Diggie">Diggie</option>
+                <option value="Dyrroth">Dyrroth</option>
+                <option value="Edith">Edith</option>
+                <option value="Esmeralda">Esmeralda</option>
+                <option value="Estes">Estes</option>
+                <option value="Eudora">Eudora</option>
+                <option value="Fanny">Fanny</option>
+                <option value="Faramis">Faramis</option>
+                <option value="Floryn">Floryn</option>
+                <option value="Freya">Freya</option>
+                <option value="Gatotkaca">Gatotkaca</option>
+                <option value="Gloo">Gloo</option>
+                <option value="Gord">Gord</option>
+                <option value="Granger">Granger</option>
+                <option value="Grock">Grock</option>
+                <option value="Guinevere">Guinevere</option>
+                <option value="Gusion">Gusion</option>
+                <option value="Hanabi">Hanabi</option>
+                <option value="Hanzo">Hanzo</option>
+                <option value="Harith">Harith</option>
+                <option value="Harley">Harley</option>
+                <option value="Hayabusa">Hayabusa</option>
+                <option value="Helcurt">Helcurt</option>
+                <option value="Hilda">Hilda</option>
+                <option value="Hylos">Hylos</option>
+                <option value="Irithel">Irithel</option>
+                <option value="Jawhead">Jawhead</option>
+                <option value="Johnson">Johnson</option>
+                <option value="Joy">Joy</option>
+                <option value="Julian">Julian</option>
+                <option value="Kadita">Kadita</option>
+                <option value="Kagura">Kagura</option>
+                <option value="Kaja">Kaja</option>
+                <option value="Karina">Karina</option>
+                <option value="Karrie">Karrie</option>
+                <option value="Khaleed">Khaleed</option>
+                <option value="Khufra">Khufra</option>
+                <option value="Kimmy">Kimmy</option>
+                <option value="Lancelot">Lancelot</option>
+                <option value="Lapu-Lapu">Lapu-Lapu</option>
+                <option value="Layla">Layla</option>
+                <option value="Leomord">Leomord</option>
+                <option value="Lesley">Lesley</option>
+                <option value="Ling">Ling</option>
+                <option value="Lolita">Lolita</option>
+                <option value="Lunox">Lunox</option>
+                <option value="Luo Yi">Luo Yi</option>
+                <option value="Lylia">Lylia</option>
+                <option value="Martis">Martis</option>
+                <option value="Mathilda">Mathilda</option>
+                <option value="Melissa">Melissa</option>
+                <option value="Minsitthar">Minsitthar</option>
+                <option value="Minotaur">Minotaur</option>
+                <option value="Moskov">Moskov</option>
+                <option value="Natalia">Natalia</option>
+                <option value="Natan">Natan</option>
+                <option value="Novaria">Novaria</option>
+                <option value="Odette">Odette</option>
+                <option value="Paquito">Paquito</option>
+                <option value="Pharsa">Pharsa</option>
+                <option value="Phoveus">Phoveus</option>
+                <option value="Popol and Kupa">Popol and Kupa</option>
+                <option value="Rafaela">Rafaela</option>
+                <option value="Roger">Roger</option>
+                <option value="Ruby">Ruby</option>
+                <option value="Saber">Saber</option>
+                <option value="Selena">Selena</option>
+                <option value="Silvanna">Silvanna</option>
+                <option value="Sun">Sun</option>
+                <option value="Terizla">Terizla</option>
+                <option value="Thamuz">Thamuz</option>
+                <option value="Tigreal">Tigreal</option>
+                <option value="Vale">Vale</option>
+                <option value="Valentina">Valentina</option>
+                <option value="Valir">Valir</option>
+                <option value="Vexana">Vexana</option>
+                <option value="Wanwan">Wanwan</option>
+                <option value="X.Borg">X.Borg</option>
+                <option value="Xavier">Xavier</option>
+                <option value="Yin">Yin</option>
+                <option value="Yi Sun-shin">Yi Sun-shin</option>
+                <option value="Yve">Yve</option>
+                <option value="Yu Zhong">Yu Zhong</option>
+                <option value="Zhask">Zhask</option>
+                <option value="Zilong">Zilong</option>
+              </select>
+            </div>
           </div>
         </div>
-    </div> 
+      </div>
+    </div>
   );
 };
